% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{codon_distance}
\alias{codon_distance}
\title{Codon distance}
\usage{
codon_distance(seq_variant1, seq_variant2, proportion = FALSE)
}
\arguments{
\item{seq_variant1}{string: coding dna,  must be in frame}

\item{seq_variant2}{string: coding dna,  must be in frame}

\item{proportion}{logical: if true, returns the distance as a proportion
(1 = all codons are different, 0 = no differences)}
}
\value{
int, number of codon diferences
}
\description{
compute the number of codon differences between the two sequences
}
\examples{
codon_distance("ATGCTG", "ATGCTT")
}
