% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codones}
\alias{codones}
\title{The genetic code
#' \describe{
  \item{codon}{three letter nucleotide codon}
}}
\format{A tibble with 64 rows and 1 variable}
\usage{
codones
}
\description{
The genetic code
#' \describe{
  \item{codon}{three letter nucleotide codon}
}
}
\keyword{datasets}
