% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_codons}
\alias{count_codons}
\title{Count codons in DNA sequence}
\usage{
count_codons(secuencia)
}
\arguments{
\item{secuencia}{character: coding dna,  must be in frame}
}
\value{
The frequency of the codons in \code{seq}
}
\description{
Counts the frequency of each triplete in sequence, assumes
the sequences is in frame
}
\examples{
count_codons("ACGGGG")
}
