% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinonimous-mutant.R
\name{evolution}
\alias{evolution}
\title{Generate synonimous mutant variants}
\usage{
evolution(secuencia, sampling_distribution, mutation_rate = 0.3,
  n_daughters = 10)
}
\arguments{
\item{secuencia}{string, coding DNA sequence in frame}

\item{sampling_distribution}{function, \code{\link{sample_synonimous_codon}}
the sampling codon distribution}

\item{mutation_rate}{number of positions to be mutated expressed as a percentage
max = 1 (mutate all codons), min = 0 (mutate no codons).
the 1st and last codon are never changed}

\item{n_daughters}{int, the number of random synonimous sequences to generate
at each function call}
}
\value{
function, sinonimous generator: A function that takes as input
a sequence and will create sinonimous mutations in that sequence
}
\description{
Generate synonimous mutant variants
}
\examples{
sampling_distribution <- sample_synonimous_codon(sampling_codon_distribution = sampling_deoptimization)
seq <- "ATGCCCGGGATGATGTTT"
evolution(seq, sampling_distribution)
}
