% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gc_amino_to_codons_l}
\alias{gc_amino_to_codons_l}
\title{Genetic code amino -> codons}
\format{
list mapping amino acid to codons
}
\usage{
gc_amino_to_codons_l
}
\description{
The genetic code
}
\keyword{datasets}
