% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gc_codons_to_amino}
\alias{gc_codons_to_amino}
\title{Genetic code codons -> amino acid}
\format{
named character vector
}
\usage{
gc_codons_to_amino
}
\description{
The genetic code
}
\keyword{datasets}
