% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_optimality}
\alias{human_optimality}
\title{Codon optimality code for human}
\format{
A data frame with 64 rows and 3 variables:
}
\usage{
human_optimality
}
\description{
The codon optimality code of human, figure 1G
and supplemental figure 2D from Medina et al 2020
\describe{
  \item{codon}{codon, DNA}
  \item{amino_acid}{amino acid, one letter amino acid code}
  \item{optimality}{optimality score values}
  ...
}
}
\keyword{datasets}
