% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrna_stability_model}
\alias{mrna_stability_model}
\title{Predictive model of mRNA stability}
\format{
An object of class \code{elnet} (inherits from \code{glmnet}) of length 12.
}
\usage{
mrna_stability_model
}
\description{
lasso (glmnet) model to predic the mRNA stability bassed on coding
sequence information
NOTE: see the file: data-raw/modeling.R where this model is trained
}
\keyword{datasets}
