% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nucleotide_distance}
\alias{nucleotide_distance}
\title{Nucleotide distance}
\usage{
nucleotide_distance(seq_variant1, seq_variant2, proportion = FALSE)
}
\arguments{
\item{seq_variant1}{string: coding dna,  must be in frame}

\item{seq_variant2}{string: coding dna,  must be in frame}

\item{proportion}{logical: if true, returns the distance as a proportion
(1 = all codons are different, 0 = no differences)}
}
\value{
int, number of codon diferences
}
\description{
compute the number of nucleotide differences between the two sequences
assumes the sequences are same length and aligned
}
\examples{
nucleotide_distance("ATGCTG", "ATGCTT")
}
