% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimizer}
\alias{optimizer}
\title{Genetic algorithm for mRNA stability optimization}
\usage{
optimizer(
  sequence_to_optimize,
  specie = "human",
  n_iterations = 100,
  make_more_optimal = T
)
}
\arguments{
\item{sequence_to_optimize}{character, coding DNA sequence from start codon to
stop codon}

\item{specie}{character, species}

\item{n_iterations}{integer, number of evolution iterations}

\item{make_more_optimal}{logical, If true the sequence is optimized, if false
is is deoptimized (less optimal)}
}
\value{
tibble data frame, the best sequence at each iteration
}
\description{
Genetic algorithm for mRNA stability optimization
}
\examples{
suppressMessages(optimizer(test_seq, n_iterations = 1))
}
