% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optipred}
\alias{optipred}
\title{Predictions for endogenous sequences}
\format{
A data frame with 34,290 rows and 3 variables:
\describe{
  \item{gene_id}{Ensembl gene id}
  \item{specie}{The specie one of: human, fish, mouse, or xenopus}
  \item{optimality}{predicted mRNA stability}
}
}
\usage{
optipred
}
\description{
NOTE: see the file: data-raw/predict_endogenous_sequences.R where i generate this data
Predicted mRNA stability for endogenous genes in each species
}
\keyword{datasets}
