% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_optimization}
\alias{plot_optimization}
\title{Plot optimization path and mRNA stability level}
\usage{
plot_optimization(optimization_run)
}
\arguments{
\item{optimization_run}{tibble: the output result of \code{\link{optimizer}}}
}
\value{

}
\description{
plots the optimization path and the mRNA stability level
}
