% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-preprocessing.R
\name{preprocess_secuences}
\alias{preprocess_secuences}
\title{Preprocess sequence for prediction mRNA stability}
\usage{
preprocess_secuences(secuencia, specie_ = "human")
}
\arguments{
\item{secuencia}{character: coding dna,  must be in frame}

\item{specie_}{character: one of human, mouse, fish, or xenopus}
}
\value{
A tibble: \code{length(secuencia)} x 423, preprocessed data for estimating
mrna stability
}
\description{
Preprocess sequence for prediction mRNA stability
}
\examples{
preprocess_secuences(test_seq, specie_ = "mouse")
}
