% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preprocessing_recipe}
\alias{preprocessing_recipe}
\title{Data pre-processing pipeline}
\format{
An object of class \code{recipe} of length 8.
}
\usage{
preprocessing_recipe
}
\description{
The pipeline to pre-process the data for predicting the mRNA stability
NOTE: see the file: data-raw/modeling.R where I generated this
}
\keyword{datasets}
