% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pretty_print_seq}
\alias{pretty_print_seq}
\title{Pretty print DNA sequence}
\usage{
pretty_print_seq(secuencia)
}
\arguments{
\item{secuencia}{character: coding dna,  must be in frame}
}
\value{

}
\description{
adds a line break every 80 characters to pretty print the secuence on screen
}
\examples{
cat(pretty_print_seq(test_seq))
}
