% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyscripts.R
\name{run_optimization_shinny}
\alias{run_optimization_shinny}
\title{Script to run the optimization in shiny app}
\usage{
run_optimization_shinny(secuencia, animal)
}
\arguments{
\item{secuencia}{character coding DNA.}

\item{animal}{character specie.}
}
\value{
tibble: results with optimizatin path
}
\description{
This optimization is run in both direction -> up (optimized) and down (deoptimized)
}
