% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampling_deoptimization}
\alias{sampling_deoptimization}
\title{Sampling probabilities (deoptimization)}
\format{
list mapping amino acid to synonimous codon sampling probabilities
}
\usage{
sampling_deoptimization
}
\description{
The genetic code
}
\keyword{datasets}
