% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampling_optimization}
\alias{sampling_optimization}
\title{Sampling probabilities (optimization)}
\format{list mapping amino acid to synonimous codon sampling probabilities}
\usage{
sampling_optimization
}
\description{
The genetic code
}
\keyword{datasets}
