% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_by_codons}
\alias{split_by_codons}
\title{Split a sequence by codons}
\usage{
split_by_codons(secuencia)
}
\arguments{
\item{secuencia}{character: coding dna,  must be in frame}
}
\value{
character vector, each element is a codon
and they come in the same order
}
\description{
This is an internal function
}
