% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing}
\alias{testing}
\title{Testung data}
\format{A data frame with 67,775 rows and 8 variables:
\describe{
  \item{gene_id}{Ensembl gene id}
  \item{specie}{The specie one of: human, fish, mouse, or xenopus}
  \item{cell_type}{The cell type: 293t, embryo mz, hela, mES cells, k562, or RPE}
  \item{datatype}{experimental technique used to obtain the mRNA stability measurements (decay rate)}
  \item{decay_rate}{The decay rate (variable to predict), scaled with respect to each experiment}
  \item{utrlenlog}{ size in log scale of the 5' UTR}
  \item{coding}{coding dna sequence in frame}
  \item{cdslenlog}{size in log scale of the coding sequence}
}}
\usage{
testing
}
\description{
This is the data used to test the predictive model
}
\details{
The decay rate measurements were obtained from the following papers:
\enumerate{
\item Bazzini, Ariel A., et al. "Codon identity regulates mRNA stability and translation efficiency during the maternal‐to‐zygotic transition."
\item Wu, Qiushuang, et al. "Translation affects mRNA stability in a codon-dependent manner in human cells." Elife 8 (2019): e45396.
\item Herzog, Veronika A., et al. "Thiol-linked alkylation of RNA to assess expression dynamics." Nature methods 14.12 (2017): 1198.
}
}
\keyword{datasets}
