% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{translate}
\alias{translate}
\title{translate DNA sequence to amino acid}
\usage{
translate(secuencia)
}
\arguments{
\item{secuencia}{string: coding dna,  must be in frame}
}
\value{
string, amino acid sequence
}
\description{
translate DNA sequence to amino acid
}
\examples{
translate("ATG")
}
