% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{unscale_decay_to_mouse}
\alias{unscale_decay_to_mouse}
\title{Convert scaled decay rate to unscaled versio (half-life)}
\usage{
unscale_decay_to_mouse(scaled_decay_rate, return_hl = TRUE)
}
\arguments{
\item{scaled_decay_rate}{double: decay rate (observed or predicted value)}

\item{return_hl}{logical: if true returns the half life else the decay rata (unscaled)}
}
\value{
half life
}
\description{
For the purpose of training the model I scaled the decay rates to std = 1
and mean = 0. This function puts the decay rate in the slam-seq mouse scale.
So I can get back a half life
}
\examples{
unscale_decay_to_mouse(-2)
}
