% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_sequence}
\alias{validate_sequence}
\title{Check correct input DNA sequence}
\usage{
validate_sequence(secuencia)
}
\arguments{
\item{secuencia}{character: coding dna,  must be in frame}
}
\value{
throws an error if the sequence contains invalid characters or is not a
multiple of 3
}
\description{
Check correct input DNA sequence
}
\examples{
validate_sequence(test_seq)
}
