% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinonimous-mutant.R
\name{variante_sinonima}
\alias{variante_sinonima}
\title{Generate synonimous mutant variant}
\usage{
variante_sinonima(secuencia, sampling_distribution, percentage = 0.3)
}
\arguments{
\item{secuencia}{string, coding DNA sequence in frame}

\item{sampling_distribution}{function, \code{\link{sample_synonimous_codon}}
the sampling codon distribution}

\item{percentage}{float, the percentage of the sequence to be mutated
max = 1 (mutate all codons), min = 0 (mutate no codons). This function wont change
the 1st and last codon}
}
\value{
string, synonimous variant sequence
}
\description{
Generate synonimous mutant variant
}
\examples{
sampling_distribution <- sample_synonimous_codon(sampling_codon_distribution = sampling_deoptimization)
seq <- "ATGCCCGGGATGATGTTT"
variante_sinonima(seq, sampling_distribution, .5)
}
