% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_evolution}
\alias{visualize_evolution}
\title{Visualiztion tools}
\usage{
visualize_evolution(optimization_run, draw_heatmap = T)
}
\arguments{
\item{optimization_run}{tibble: the output result of \code{\link{optimizer}}}

\item{draw_heatmap}{logical: If true draws a heatmap, otherwise draws ... TODO}
}
\value{
plot: ggplot2 object
}
\description{
Helper function to visualize the optimization
}
