% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyscripts.R
\name{viz_result_shiny}
\alias{viz_result_shiny}
\title{Visualize optimization results in shiny}
\usage{
viz_result_shiny(result, animal)
}
\arguments{
\item{result}{tibble output of \code{\link{run_optimization_shinny}}}

\item{animal}{string one of (fish, human, mouse, or xenopus)}
}
\value{
ggplot object
}
\description{
Plots the optimization result to be displaye in the shiny appo
}
