% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimizer}
\alias{optimizer}
\title{Genetic algorithm for mRNA stability optimization}
\usage{
optimizer(
  sequence_to_optimize,
  specie = "human",
  n_iterations = 15,
  make_more_optimal = T,
  mutation_Rate = 0.4,
  max_abs_val = Inf,
  n_Daughters = 3
)
}
\arguments{
\item{sequence_to_optimize}{character, coding DNA sequence from start codon to
stop codon}

\item{specie}{character, species}

\item{n_iterations}{integer, number of evolution iterations}

\item{make_more_optimal}{logical, If true the sequence is optimized, if false}

\item{mutation_Rate}{doblue, number of mutation to introduce at eact iteration
given as a probability (1 = max (mutate all positions), 0 = min (no mutations))}

\item{max_abs_val}{A maximum value (absolute value) such that the sequences
cannot go beyond that value. Infinite is used as default, in this case there is not
a limit for the sequences}

\item{n_Daughters}{integer, number of child sequences to explore at each iteration
is is deoptimized (less optimal)}
}
\value{
a [tibble][tibble::tibble-package], the best sequence at each iteration
}
\description{
Genetic algorithm for mRNA stability optimization
}
\examples{
suppressMessages(optimizer(test_seq, n_iterations = 1))
}
