% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-stability.R
\name{predict_stability}
\alias{predict_stability}
\title{Predict mRNA stability functional factory}
\usage{
predict_stability(specie)
}
\arguments{
\item{specie}{character: one of human, mouse, fish, or xenopus}
}
\value{

}
\description{
This function will generate a function that predicts the mRNA stability of the
given species in the input sequences The prediction is generated with the model:
\code{mrna_stability_model}. The output function will take as input a set of
dna sequences (vector) these sequences should not contain repeated elements
}
\examples{
predictor_fish <- predict_stability("fish")
some_seqs <- testing$coding[1:3]
predicted_stability <- predictor_fish(some_seqs)
}
