% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-preprocessing.R
\name{preprocess_secuences}
\alias{preprocess_secuences}
\title{Preprocess sequence for prediction mRNA stability}
\usage{
preprocess_secuences(secuencias, specie_ = "human")
}
\arguments{
\item{secuencias}{character, a vector of dna sequences to predict. If
more than one sequence is supplied, then the sequences should be unique. (No
repeated sequences)}

\item{specie_}{character: one of human, mouse, fish, or xenopus}
}
\value{
A tibble: \code{length(secuencia)} x 423, preprocessed data for estimating
mrna stability
}
\description{
Preprocess sequence for prediction mRNA stability
}
\examples{
preprocess_secuences(test_seq, "mouse")
}
